refactor: Shift to `storage.MediTrackerFileConfig`
